pyproject.toml
[project]
name = "sglang-service"
version = "0.1.0"
description = "SGLang custom generation service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "blobfile==3.0.0",
    "build==1.4.0",
    "compressed-tensors==0.14.0.1",
    "datasets==4.7.0",
    "einops==0.8.2",
    "fastapi==0.135.1",
    "hf-transfer==0.1.9",
    "huggingface-hub==0.36.2",
    "interegular==0.3.3",
    "llguidance==0.7.30",
    "modelscope==1.35.0",
    "msgspec==0.20.0",
    "ninja==1.13.0",
    "openai==2.6.1",
    "openai-harmony==0.0.4",
    "orjson==3.11.7",
    "outlines==0.1.11",
    "packaging==26.0",
    "partial-json-parser==0.2.1.1.post7",
    "pillow==12.1.1",
    "prometheus-client==0.24.1",
    "psutil==7.2.2",
    "pybase64==1.4.3",
    "pydantic==2.12.5",
    "pynvml==13.0.1",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.22",
    "pyzmq==27.1.0",
    "scipy==1.17.1",
    "sentencepiece==0.2.1",
    "sglang==0.5.9",
    "soundfile==0.13.1",
    "tiktoken==0.12.0",
    "timm==1.0.16",
    "torch==2.9.1",
    "torchao==0.9.0",
    "transformers==4.57.1",
    "uvicorn[standard]==0.41.0",
    "uvloop==0.22.1",
    "xgrammar==0.1.27",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
